@@ -4,7 +4,7 @@
   "description": "A hapi view engine for React components.",
   "main": "index.js",
   "scripts": {
-    "test": "lab -t 100 -c -L -I 'Reflect,core,_babelPolyfill,regeneratorRuntime,__core-js_shared__'",
+    "test": "lab -t 100 -c -L -I '__core-js_shared__'",
     "test-cover": "lab -t 100 -c -r html -o test/artifacts/coverage.html && open test/artifacts/coverage.html",
     "simple-example": "node examples/simple/server",
     "layout-example": "node examples/layout/server",
@@ -38,16 +38,16 @@
   "devDependencies": {
     "babel": "6.x.x",
     "babel-core": "6.x.x",
-    "babel-loader": "6.x.x",
+    "babel-loader": "7.x.x",
     "babel-preset-es2015": "6.x.x",
     "babel-preset-react": "6.x.x",
     "code": "4.x.x",
-    "hapi": "15.x.x",
+    "hapi": "16.x.x",
     "inert": "4.x.x",
-    "lab": "11.x.x",
+    "lab": "13.x.x",
     "react": "15.x.x",
     "react-dom": "15.x.x",
     "vision": "4.x.x",
-    "webpack": "1.x.x"
+    "webpack": "2.x.x"
   }
 }
