@@ -1,15 +1,14 @@
 const React = require('react');
-const Layout = require('./layout.jsx');
 
 
-const Component = React.createClass({
-    render: function () {
+class AboutView extends React.Component {
+    render () {
 
         return (
             <h1>About the plot device.</h1>
         );
     }
-});
+}
 
 
-module.exports = Component;
+module.exports = AboutView;
