@@ -1,20 +1,21 @@
 const React = require('react');
 
 
-const Component = React.createClass({
-    handleClick: function () {
+class App extends React.Component {
+    handleClick () {
       alert('Hi ' + this.props.foo);
-    },
-    render: function () {
+    }
+
+    render () {
 
         return (
           <div>
               <h1>Foo: ({this.props.foo})</h1>
-              <button onClick={this.handleClick}>Event test</button>
+              <button onClick={this.handleClick.bind(this)}>Event test</button>
           </div>
         );
     }
-});
+}
 
 
-module.exports = Component;
+module.exports = App;
