@@ -1,8 +1,8 @@
 const React = require('react');
 
 
-const Component = React.createClass({
-    render: function () {
+class Component extends React.Component {
+    render () {
 
         return (
             <html>
@@ -17,7 +17,7 @@ const Component = React.createClass({
             </html>
         );
     }
-});
+}
 
 
 module.exports = Component;
