@@ -1,14 +1,14 @@
 const React = require('react');
 
 
-const Component = React.createClass({
-    render: function () {
+class Component extends React.Component {
+    render () {
 
         return (
             <div>Navbar</div>
         );
     }
-});
+}
 
 
 module.exports = Component;
