@@ -12,7 +12,7 @@ require('babel-core/register')({
 
 
 const server = new Hapi.Server();
-server.connection();
+server.connection({ port: process.env.PORT });
 server.register([Inert, Vision], (err) => {
 
     if (err) {
